fix: Move `find_first` into `collections` folder
The agent created it in a `collections.py` file instead, which then overrode any reference to the folder.
