+++ src/dataset_foundry/utils/collections/find_first.py
